+++ client/CHANGELOG.md
@@ -1,7 +1,12 @@
-## 17.10
+## 17.11
 
 * Bump CAPI models to 15.10.0
   * updates `aliasPaths` definition to include `ceasedToBeCanonicalAt` datetime
+
+## 17.10
+
+* Add `Format` to `CapiModelEnrichment`.
+  * This adds the `Design`, `Display` and `Theme` traits to be consumed by platforms.
 
 ## 17.9